requirements.txt
torch==2.1.0
transformers==4.36.2
accelerate==0.24.1
datasets==2.14.6
peft==0.6.2
bitsandbytes==0.41.2.post2
trl==0.7.4
sentence-transformers==2.2.2
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
tqdm==4.66.1
scikit-learn==1.3.2
scipy==1.11.4
pandas==2.1.3
pyarrow==14.0.1
requests==2.31.0
pyyaml==6.0.1
regex==2023.10.3
packaging==23.2
safetensors==0.4.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
pgvector==0.2.4